Relocate jsonmessages
